% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restrict.R
\name{restrict}
\alias{restrict}
\title{Extract data from netcdf file based on land id}
\usage{
restrict(f, var_id, var_land_id = "land_id", ids, par = FALSE)
}
\arguments{
\item{f}{the path to the netcdf}

\item{var_id}{variable to extract}

\item{var_land_id}{variable used for land ids in the netcdf file}

\item{ids}{a vector of land ids}

\item{par}{run in parallel mode (not implemented)}
}
\value{
a data.table
}
\description{
(Relatively) fast way to extract data from netcdf file structured
  by land id. The use case is to extract a subset of the full
  grid, based on a set of land ids. These can be non-consecutive.
}
